Pequenia correction en tipo de dato (xcoord y ycoord) y en formato de impresion /t a \t

+++ src/model/vo/VOMovingViolation.java
@@ -145,12 +145,12 @@ public class VOMovingViolation {
 		return Integer.parseInt(FINEAMT);
 	}
 	
-	public int getXCoord() {
-		return Integer.parseInt(XCOORD);
+	public float getXCoord() {
+		return Float.parseFloat(XCOORD);
 	}
 	
-	public int getYCoord() {
-		return Integer.parseInt(YCOORD);
+	public float getYCoord() {
+		return Float.parseFloat(YCOORD);
 	}
